Updating index for friends
My index is a bit more complete and has much more information than most people care about (as well as a few projects that don't appear here, like the Big Data/NLP project which doesn't make much sense as a github project at the moment). Honestly it's closer to a diary or journal than an index. I've also been considering restructuring my naming scheme to organize projects a bit better. This is me stripping down to the bare-bones description with maybe an insight or two for each project as it stands on github.